move code into new class SamplePackBuilder
